Update project.tex
I added two definitions and one reference. In order to expand the table,
we need the derived category context, because in general the exceptional
inverse image is defined only at level of derived categories. The two exception
are obviously closed immersion (already in the table), and the open immersion
case.